requirements.txt
# ─────────────────────────────────────────────────────────────
# A股量化系统 Python 依赖
# ─────────────────────────────────────────────────────────────

# ── 数据源 ───────────────────────────────────────────────────
tushare>=1.4.0

# ── 计算引擎 ─────────────────────────────────────────────────
polars>=0.20.0
numpy>=1.26.0
pandas>=2.1.0          # Qlib 内部仍依赖 pandas

# ── 数据库 ───────────────────────────────────────────────────
psycopg2-binary>=2.9.9     # PostgreSQL 驱动
sqlalchemy>=2.0.0           # ORM / 连接池
connectorx>=0.4.2           # Polars read_database_uri 后端
alembic>=1.13.0             # 数据库迁移

# ── 缓存 / 任务队列 ──────────────────────────────────────────
redis>=5.0.0
celery>=5.3.0               # 异步任务（可选）

# ── 策略框架 ─────────────────────────────────────────────────
# Qlib 需要单独通过 git 安装最新版，或使用以下版本
pyqlib>=0.9.0

# ── 量化工具 ─────────────────────────────────────────────────
ta>=0.11.0                  # 技术指标（纯 Python 实现，兼容 pandas 2.x）
# 若要使用原版 ta-lib，需在 Dockerfile 中先安装 C 库:
# apt-get install libta-lib-dev && pip install TA-Lib

# ── 可视化 ───────────────────────────────────────────────────
matplotlib>=3.8.0
plotly>=5.18.0
pyecharts>=2.0.0            # ECharts Python 封装

# ── JupyterLab ───────────────────────────────────────────────
jupyterlab>=4.0.0
ipykernel>=6.29.0
ipywidgets>=8.1.0

# ── 数据验证 / 配置 ──────────────────────────────────────────
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# ── 工具 ─────────────────────────────────────────────────────
loguru>=0.7.0               # 日志
tqdm>=4.66.0                # 进度条
schedule>=1.2.0             # 定时任务
httpx>=0.26.0               # 异步 HTTP
tenacity>=8.2.0             # 重试
pyarrow>=14.0.0             # Parquet / Arrow 格式（Polars 需要）
